+++ Makefile
@@ -1,12 +1,12 @@
 # **************************************************************************** #
 #                                                                              #
-#                                                         :::      ::::::::    #
-#    Makefile                                           :+:      :+:    :+:    #
-#                                                     +:+ +:+         +:+      #
-#    By: qbeukelm <qbeukelm@student.42.fr>          +#+  +:+       +#+         #
-#                                                 +#+#+#+#+#+   +#+            #
-#    Created: 2023/12/03 13:06:57 by quentinbeuk       #+#    #+#              #
-#    Updated: 2024/01/06 17:04:29 by qbeukelm         ###   ########.fr        #
+#                                                         ::::::::             #
+#    Makefile                                           :+:    :+:             #
+#                                                      +:+                     #
+#    By: qbeukelm <qbeukelm@student.42.fr>            +#+                      #
+#                                                    +#+                       #
+#    Created: 2023/12/03 13:06:57 by quentinbeuk   #+#    #+#                  #
+#    Updated: 2024/01/07 13:09:40 by quentinbeuk   ########   odam.nl          #
 #                                                                              #
 # **************************************************************************** #
 
@@ -31,28 +31,36 @@ RESET				:= \033[0m
 
 
 # ===== Sources =====
-SOURCES 			= minishell.c \
-						shell_init.c \
-						list.c 
+SOURCES 				= minishell.c \
+							shell_init.c 
 
-SOURCES_LEXER 		= lexer.c \
-						quote.c \
-						quote_matcher.c \
-						split.c \
-						tokenize_quotes.c
+SOURCES_LEXER 			= lexer.c \
+							quote.c \
+							quote_matcher.c \
+							tokenize_quotes.c
+
+SOURCES_LEXER_SPLIT	 	= split.c \
+							allocate_strings.c
+
+SOURCES_UTILS			= clean_exit.c \
+							list.c
 
 
 # ===== Manage Directories =====
-INC 				= inc
-DIR_OBJ				= obj
+INC 					= inc
+DIR_OBJ					= obj
 
-DIR_SOURCES 		= sources
-DIR_SOURCES_LEXER	= sources/lexer
+DIR_SOURCES 			= sources
+DIR_SOURCES_LEXER		= sources/lexer
+DIR_SOURCES_LEXER_SPLIT = sources/lexer/split
+DIR_SOURCES_UTILS		= sources/utils
 
 
 # ===== Object Files =====
 OBJ = $(addprefix $(DIR_OBJ)/, $(SOURCES:.c=.o)) \
-	$(addprefix $(DIR_OBJ)/, $(SOURCES_LEXER:.c=.o))
+	$(addprefix $(DIR_OBJ)/, $(SOURCES_LEXER:.c=.o)) \
+	$(addprefix $(DIR_OBJ)/, $(SOURCES_LEXER_SPLIT:.c=.o)) \
+	$(addprefix $(DIR_OBJ)/, $(SOURCES_UTILS:.c=.o))
 
 
 # ===== Rules =====
@@ -69,6 +77,12 @@ $(DIR_OBJ)/%.o: $(DIR_SOURCES)/%.c | $(DIR_OBJ)
 	@$(CC) $(CFLAGS) -I$(INC) -c $< -o $@
 
 $(DIR_OBJ)/%.o: $(DIR_SOURCES_LEXER)/%.c | $(DIR_OBJ)
+	@$(CC) $(CFLAGS) -I$(INC) -c $< -o $@
+
+$(DIR_OBJ)/%.o: $(DIR_SOURCES_LEXER_SPLIT)/%.c | $(DIR_OBJ)
+	@$(CC) $(CFLAGS) -I$(INC) -c $< -o $@
+
+$(DIR_OBJ)/%.o: $(DIR_SOURCES_UTILS)/%.c | $(DIR_OBJ)
 	@$(CC) $(CFLAGS) -I$(INC) -c $< -o $@
 
 $(DIR_OBJ):
