+++ sources/minishell.c
@@ -1,63 +1,43 @@
 /* ************************************************************************** */
 /*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   minishell.c                                        :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: qbeukelm <qbeukelm@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/12/03 13:13:49 by quentinbeuk       #+#    #+#             */
-/*   Updated: 2024/01/06 17:03:43 by qbeukelm         ###   ########.fr       */
+/*                                                        ::::::::            */
+/*   minishell.c                                        :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: qbeukelm <qbeukelm@student.42.fr>            +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2023/12/03 13:13:49 by quentinbeuk   #+#    #+#                 */
+/*   Updated: 2024/01/07 13:07:35 by quentinbeuk   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-
-static char	*retrieve_command(void)
+static bool retrieve_command(t_shell *shell)
 {
 	char *command;
 
-	command = readline("[minishell]: ");
+	command = readline(CYELLOW "[minishell]: " RESET_COLOR);
 	if (command == NULL)
 	{
-		// TODO Free
+		// TODO clean_exit()
+		return (FAILURE);
 	}
-	return (command);
-}
-
-void	print_token(t_token *tokens)
-{
-	printf("\n\n========print tokens========\n");
-	while (tokens)
-	{
-		printf("%s \n", tokens->value);
-		tokens = tokens->next;
-	}
-}
-
-static void	finish_command(t_shell *shell)
-{
-	shell->cmd->single_quote = 0;
-	shell->cmd->double_quote = 0;
+	save_command(command, shell);
+	return (SUCCESS);
 }
 
 static bool	run(t_shell *shell)
 {
-	char	*command;
-	t_token	*tokens_head;
-
-	tokens_head = NULL;
 	while (1) 
 	{
-		command = retrieve_command();
-		input_init(command, shell); // <- Move
-		tokens_head = tokens_builder_manager(shell);
+		retrieve_command(shell);
+		tokens_builder_manager(shell);
 		finish_command(shell);
 		print_token(shell->tokens);
 
 		printf("\n--------------------End--------------------\n\n");
 	}
-	return (true);
+	return (SUCCESS); // <- Return status to main
 }
 
 int	main(int argc, char **argv, char **envp)
@@ -66,11 +46,8 @@ int	main(int argc, char **argv, char **envp)
 
 	shell = shell_init();
 	run(shell);
-
-	// print uncsured param
+	
 	printf("argc: %d, %p, %p", argc, argv, envp);
-	// init data
-	// Store params
-	// run
+	
 	return (0);
 }
