+++ sources/utils/list.c
@@ -1,16 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   list.c                                             :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: qbeukelm <qbeukelm@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/12/22 15:12:33 by qbeukelm          #+#    #+#             */
-/*   Updated: 2024/01/06 16:59:50 by qbeukelm         ###   ########.fr       */
+/*                                                        ::::::::            */
+/*   list.c                                             :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: qbeukelm <qbeukelm@student.42.fr>            +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2023/12/22 15:12:33 by qbeukelm      #+#    #+#                 */
+/*   Updated: 2024/01/07 13:08:45 by quentinbeuk   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
+#include "../../includes/minishell.h"
 
 t_token	*lstlast(t_token *token)
 {
